Add template email on sheriff notifier
We need this template for a more informative message when we notify the sheriffs.

While the templates seem to not work in another project we hope to be able to collect some more debuging info by adding it now. See https://crbug.com/1231981#c9 .

Bug: chromium:1223045
Change-Id: Iba1cb3b3a23c0c06790ded394e86c7dd09f4d1ea
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3104548
Reviewed-by: Michael Achenbach <machenbach@chromium.org>
Commit-Queue: Liviu Rau <liviurau@chromium.org>