[CSS]: Edge cases in panels
Some uses of registerRequiredCSS in panels/network, panels/accessibility
and panels/timeline do not use this.registerRequiredCSS. Instead they
register CSS files on properties, such as a treeOutline. These cases
were not dealt with by the script as you could not add a wasShown
method where the CSS files could be registered. Instead the changes
for these have to be done in the same place, replacing
registerRequiredCSS with registerCSSFiles.

Bug: 1106746
Change-Id: I6d6c29817631de6e687ec84466b9ce275fac98e3
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3103792
Commit-Queue: Kriti Sapra <kritisapra@google.com>
Reviewed-by: Tim van der Lippe <tvanderlippe@chromium.org>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>