Factor out device size input elements into custom elements
Now, we can use an inline style tag, rather than relying on the
globally available legacy stylesheet (or at least parts of it).

Eventually, we can do this for the other parts of the toolbar and
at that point we can remove the dependency on the full ui legacy
system.

R=jacktfranklin@chromium.org,kritisapra@google.com

Bug: 1241064
Change-Id: Ia16801f247a637572eb6d552074d2f55ff70e4b0
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3103794
Commit-Queue: Tim van der Lippe <tvanderlippe@chromium.org>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Kriti Sapra <kritisapra@google.com>