Heuristically identify identifier types in CDP
This CL uses a heuristic to classify identifier types:
If the type ends with Id or ID it is considered an
identifier type. The CL also introduces an override
in protocol_dts_generator.ts which can be used to
add more types, or disable the heuristic for specific
types.

DISABLE_THIRD_PARTY_CHECK=disable

Fixed: chromium:1226471
Change-Id: I4696463622ef026c48e1abeefd48aa438f9d11f6
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3110425
Commit-Queue: Sigurd Schneider <sigurds@chromium.org>
Reviewed-by: Tim van der Lippe <tvanderlippe@chromium.org>