Allow context menu to be created with an object
I'd like to add a new option to the config menu but doing so when each
argument is passed one by one (and three are optional!) is hard. This CL
therefore adds the ability to pass in an object of options, but to avoid
a breaking change, supports both the old and the new way.

Over time I will migrate the 40+ instantiations of the context menu to
the new approach, but this will enable me to add the new option now and
unblock that work whilst I do it.

Bug: 1242389
Change-Id: Ib800d463bcd5cd02496dbea5d8124aaa97c77fa2
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3110375
Reviewed-by: Tim van der Lippe <tvanderlippe@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>