Make ids required for TreeOutline nodes + use id for `expanded` state
In the accessibility tree, we update nodes by replacing the old node
with a new node with the same node id. This change forces tree nodes to
have an id, and uses the id for tracking which nodes are currently
expanded. This ensures that we can track which nodes are expanded
consistently even when nodes are replaced.

Bug: 1226486
Change-Id: I5438bca48d87cc223ee6bf030173ed5dd2fe2851
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3115971
Commit-Queue: Johan Bay <jobay@chromium.org>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>