Fix white background being displayed for scroll bars on hover in dark mode
In bug 1244562, we've switched to using the native look and feel for
scrollbars across all panels. In some situations (on hover, on macOS),
Blink uses a heuristic based on the overflowing container's background
color to determine how the scrollbars are colored. When overflowing
elements don't have a specific background color set, Blink colors the
scrollbars using light colors, instead of analyzing the underlying
layers. This CL adds appropriate background colors to such elements, to
help Blink pick the right color.

Bug: 1245501
Change-Id: I2998403bf700e7013d0b8b9d5e05858162bb56d3
Signed-off-by: Victor Porof <victorporof@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3136215
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>