Fix size input elements in dark mode
In dark mode, when there is a pre-selected device size, we disable
the corresponding input elements. However, the styling was making
it not ideal in dark mode.

Instead, we should show the size as normal, but disable test selection.
That way, it is still readable, but also clear to the user that they
can't change it.

R=jacktfranklin@chromium.org

Fixed: 1246346
Change-Id: I3166b3141b005915b75be5af54787f911b60c1ef
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3142421
Auto-Submit: Tim van der Lippe <tvanderlippe@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>