[cleanup] Use getAsFile instead of webkitGetAsEntry
TypeScript 4.4.2 introduces types for `webkitGetAsEntry`. To use it
properly though, we would have to introduce a manual cast to the
"FileEntry" sub class.

We opted instead of replacing the usage of `getAsFile` where
applicable. In most cases, we are only interested in the file of
a drag event anyway, so no need to go via the indirection of
`webkitGetAsEntry` when we can have the file directly.

R=tvanderlippe@chromium.org

Bug: 1247593
Change-Id: Icd4309fcd1f36a2f5775cdc732091beb60b69698
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3147830
Commit-Queue: Simon Zünd <szuend@chromium.org>
Reviewed-by: Tim van der Lippe <tvanderlippe@chromium.org>