Remove special-casing for utils
Historically, we special-cased importing the legacy utils, as
Rollup would eagerly roll them up in a single bundle. However,
since then we changed the Rollup heuristic to always assume that
a different folder is a different entrypoint, including sub-folders.
As such, we are no longer including entrypoints of sub-folders in
parent folders.

However, utils was special-cased and was importing the direct files.
This normally isn't a problem, except for the fact that the
devtools_entrypoint of `ui/legacy:bundle` would now bundle the sources
of `ui/legacy/utils`, but wasn't rebuilding when it needed to.

Therefore, change the utils to a proper sub-folder and avoid the
special-casing. Since we no longer have a circular dependency
between the utils and `ui/legacy`, we can safely make this change.

The ESLint rule has also been updated to make sure we don't regress
in this area again.

R=victorporof@chromium.org

Fixed: 1148274
Change-Id: I020622c5790041d5f676bee2ef883ff0ece76695
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3148370
Auto-Submit: Tim van der Lippe <tvanderlippe@chromium.org>
Commit-Queue: Victor Porof <victorporof@chromium.org>
Reviewed-by: Victor Porof <victorporof@chromium.org>