Add `registerCSSFiles` to Toolbar
The history of traces dropdown in the performance panel unfortunately
requires injecting some styles into it for the dropdown. To enable us to
remove the appendStyle call I've therefore added registerCSSFiles to the
Toolbar, and we can inject the styles in when the button is added to the
toolbar.

In the future we hope this can be tidied up when it comes to building a
new set of generic components that will include a nicely styled
dropdown, but for now this allows us to get closer to removing
appendStyle entirely.

Bug: 1106746
Change-Id: I94b4da4d7a0fbd49e13ea6caecb552cf90708f7d
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3151805
Reviewed-by: Tim van der Lippe <tvanderlippe@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>