Update feature check for window segments
With crrev.com/c/3153295, the `getWindowSegments` method on Window is
being removed and is moving to a property on VisualViewport, in response
to feedback from the standardization process. The method was being used
as feature detection for the dual screen device emulator, so move to the
new check instead.

Bug: none
Change-Id: Ie36ee8ec6dc280da58e17a2bf7f845e47497c45b
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3160316
Reviewed-by: Paul Lewis <aerotwist@chromium.org>
Commit-Queue: Daniel Libby <dlibby@microsoft.com>