Add mixin for event handling
This CL adds a mixin based on ObjectWrapper. Its intended use is for
deep inheritance hierarchies where different subclasses want to
offer different sets of Events. The only use case in DevTools is
the legacy Widget system.

As the main use case for event handling remains to inherit from
ObjectWrapper, we decided to implement the mixin in terms of
ObjectWrapper and not the other way around. This means the
mixin is very slightly more expensive to use, due to the indirection
via a private field.

The private field on the mixin is necessary as TypeScript doesn't
allow private/protected properties on mixin classes, making this
a perfect candidate for using private fields, as #events should
never be accessed directly.

Bug: 1228674
Change-Id: I0d8295ae8c8a2ac7978b4ef3b3fae31d023bb8c4
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3159999
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Philip Pfaffe <pfaffe@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>