Increase z-index value of split widget resizer
The split widget resizer is used to resize DevTools when it is docked to
the inspected window. As such, it should overlap any content shown in
DevTools around the boders so that the docked app can always be resized.
However, the Settings screen, which makes use of the `GlassPane` utility
element is set to have an z-index of 3000 and was recently changed to
occupy the whole DevTools app (in https://crrev.com/c/3001656), which
means it overlaps the resizer that has a z-index of 500. This made it
impossible to resize DevTools when it is docked to the window and when
the Settings screen is shown.

To fix this, this CL increases the value of the resizer's z-index to
make sure it always overlaps the settings screen to keep DevTools
resizable when it is shown.

Fixed: 1248009
Change-Id: I41b7031a0da9cb1040fa1956b118f802bcd21e50
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3163137
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Andres Olivares <andoli@chromium.org>