[persistence] update active state when targets attach
When overrides are enabled on DevTools startup, the necessary
interceptors aren't installed. This is because interception is only
activated if a main target exists, which is not the case on startup.
This CL turns the NetworkPersistenceManager into a target observer,
rechecking for a main target whenever targets attach or detach.

Fixed: 1239919
Change-Id: Idb7f2a4ca14846fbd6a5d4062aad51acd7af3244
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3165098
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Philip Pfaffe <pfaffe@chromium.org>