Do not update FrameDetailsView whenever a resource is added
Even though redrawing of the FrameDetailsView is throttled, there are
situations where FrameDetailsView is being re-rendered multiple times
in short order resulting visible flickering of that panel.

Instead of increasing the throttler's timeout, I am removing the call
to re-render the FrameDetailsView whenever a resource is added to the
frame because these re-renders are not needed.

Fixed: 1247122
Change-Id: Iabad73e884a884cf3b523a54b34f3b35552d40e1
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3176140
Reviewed-by: Simon Zünd <szuend@chromium.org>
Commit-Queue: Wolfgang Beyer <wolfi@chromium.org>