Enable CPU throttling on interaction tests
This CL moves the throttling code from being specific to e2e tests to
being generic in Conductor such that you can run interaction tests with
the STRESS flag and have the CPU throttled.

Doing this causes some of the interaction tests to timeout so I
increased the timeout to a more reasonable number.

Bug: none
Change-Id: Ia7fdeba6490c62b591b0feefad885df3f999cb37
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3201979
Reviewed-by: Tim van der Lippe <tvanderlippe@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>