[DevTools Quick Open] Fix the bug that '?' can't be typed in textPrompt
There is no visible UI change, but before when user type'?', it will be caught by ShortcutRegistry.
So the solution is to change the input area to <input>.
And then use an <span> to be hided under the <input>, and to display the suggestion text.

Before: https://imgur.com/a/F2juZGn
After: https://imgur.com/a/mpt22E1

Bug: 1201997
Change-Id: I5970c680aa8099b7175562ecd27e8f34b8e2626e
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3202232
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Nancy Li <nancyly@chromium.org>