[sync] Add 'register' method to SettingsBackingStore interface
This CL adds a 'register' method to the backing store interface that
mirrors the `registerPreference` host bindings. For local and session
storage, 'register' is a no-op.

The CL also adds a unit test to ensure that `register` is always
called when settings are created.

R=dsv@chromium.org

Bug: 1245541
Change-Id: Iaf93171288ca99d939ab4562a91ee7bd44a61203
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3207867
Commit-Queue: Simon Zünd <szuend@chromium.org>
Reviewed-by: Danil Somsikov <dsv@chromium.org>