Add SettingCheckbox component
This CL adds a small web component that binds a checkbox to a boolean
setting. Similar to the legacy element, the binding of the new
component works both ways. Updating the setting updates the comopnent
and clicking the component updates the setting.

The CL also switches the settings UI to use SettingCheckbox.
Please note that the old element is still widely used e.g. in
toolbars or inline configuration of the console.

R=alexrudenko@chromium.org

Bug: 1245541
Change-Id: I256315e0c693edd9d5d3ecf081ff478c78b6badc
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3229277
Commit-Queue: Simon Zünd <szuend@chromium.org>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>