Add shortcut to search panel for expanding/collapsing all files
Originally, the feature request is only for expanding all the
files. Since the implementation is trivial and the shortcut is
already ready for collapsing functionality, I decided to add
collapsing as well.

I have chosen "Ctrl + Shift + }" for expanding and
"Ctrl + Shift + {" for collapsing.

Bug: chromium:1255073
Change-Id: I4e8ff3cf82b532346acb1ac6c247f0f361477287
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3212515
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>
Commit-Queue: Benedikt Meurer <bmeurer@chromium.org>