[sync] Add custom rendering for the 'Sync' section in the settings UI
This CL adds a new web component 'SyncSection' that shows the avatar
plus email of the currently logged in user, or a warning if the
user does not have sync enabled.

Screenshots: https://imgur.com/a/PI5I1mh

Drive-by: 'arePreferencesSynced' is only optionally provided by the
backend. This CL updates the SyncInformation type to reflect that.

R=alexrudenko@chromium.org, jacktfranklin@chromium.org

Bug: 1245541
Change-Id: Ieee32d961c4d09c64e4fc106c499e69d254c38df
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3234188
Commit-Queue: Simon Zünd <szuend@chromium.org>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>