TabbedPane#closeTabs should do nothing when no tabs are to be closed
This CL fixes a bug where the `EditBreakpointDialog` would
immediately close when a live/watch expression is active. The reason
is that each watch expression uses a distinct sourceURL but when
evaluated by V8 a script is created for each evaluation. This results
in the UISourceCode deduplication logic to run, once a single
evaluation script is parsed. "old scripts"
(aka previous invocations) are to be removed and the Sources View
tries to close the tabs of all the scripts that are no longer available.
This in turn triggers a focus update on the sources view, closing
the `EditBreakpointDialog` pre-maturely.

The fix in this case is rather straight-forward. When we request
a TabbedPane to close tabs, it should do nothing when the list of
"tabs to be closed" is empty.

R=bmeurer@chromium.org

Fixed: 1244580
Change-Id: Icc7ff0819ba37f56bb608ac0aadefc08119e060c
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3245374
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>