[Lighthouse] Recover from fatal Lighthouse errors.
A previous CL[1] refactored this file, transforming some promise
chaining into a try/catch. However, the final statement that
actually kicks off Lighthouse was not await'd, so any fatal
error is not handled as expected.

Additionally, the same CL dropped the 'close()' no-op method on the
Lighthouse port. This doesn't result in a user-facing error because
Lighthouse dismisses any error that happens during the disconnect
phase, but there is still an error message logged to console, which
is misleading as there really isn't any harm done.

[1] https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2697187
[2] https://github.com/GoogleChrome/lighthouse/issues/13273#issuecomment-958162407

Bug: 968639
Change-Id: Icf41f3ed8c3cd993f1d9e4748b8aff394c0582a5
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3258498
Commit-Queue: Connor Clark <cjamcl@chromium.org>
Reviewed-by: Paul Irish <paulirish@chromium.org>