Use observer for obtaining target in CSSOverview
By removing the module.json file (and the corresponding promise to load
it in `css_overview-meta.ts`), the timing for the panel changed. As
such, it executes its constructor earlier than before. Normally, this
isn't a problem, but in this case, the constructor was requesting the
main target from the TargetManager. With the new timing, the main target
isn't ready yet.

Instead, we should be using the TargetObserver to obtain the correct
target when it is ready. We can then propagate the relevant models to
the other views, which will use it when rendering. Since these views
render at a later stage, the models will be available in time.

R=alexrudenko@chromium.org

Fixed: 1266287
Change-Id: I83d8308576b72fb8566e401be81771c7c9ae4dd7
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3259087
Commit-Queue: Tim van der Lippe <tvanderlippe@chromium.org>
Commit-Queue: Alex Rudenko <alexrudenko@chromium.org>
Auto-Submit: Tim van der Lippe <tvanderlippe@chromium.org>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>