Restore CSS stylesheets on bfcache navigations
The order of events on bfcache navigations is different.
Therefore, this CL changes the behavior of CSSModel
to reload the agent to fetch the latest state about
CSS stylesheets when a bfcache navigation happens.

Drive-by: made the props on CSSModel more readable.
Drive-by 2: it looks like font faces were never reset on navigation
so added it.
Fixed: 1205309
Change-Id: I894a30569dea63d5745bdf9a40ec5000b98ea674
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3263398
Reviewed-by: Simon Zünd <szuend@chromium.org>
Commit-Queue: Alex Rudenko <alexrudenko@chromium.org>