Always reveal Sources panel with quick open menu
With crrev.com/c/2960171 the sources panel was set to be opened when a
breakpoint was hit only if an experiment was enabled. However this
behaviour was extended to other triggers of the sources panel, like when
a file is selected using the quick open menu. This means that using the
quick open menu to open a file in the sources panel would only work if
the experiment discussed before is enabled.

To fix this, the check for the experiment value was moved to the call
sites triggered specifically after a Debugger.paused event

Bug: none
Change-Id: I9eb53091b04e63627813e14a8ed6bc6605e92b3b
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3271700
Reviewed-by: Kim-Anh Tran <kimanh@chromium.org>
Commit-Queue: Andres Olivares <andoli@chromium.org>