Fix layout shift clustering calculations
This CL fixes an issue with how we were clustering the layout shifts (as
per https://web.dev/cls/#measure-cls-in-javascript). Because the code
looped through and assigned shifts to a cluster only when it found an
event in a new cluster, this meant that the final clustering of layout
shifts wouldn't be correctly assigned, and they would have no
_current_cluster_id, causing the UI to render incorrectly.

We still need to render _current_cluster_id conditionally (done in
https://crrev.com/c/3295534), because sometimes layout shifts cannot be
assigned to a cluster (if they have recent user input), but this CL
fixes cases where we would incorrectly not assign a cluster.

I also pulled the method out so we can unit test it.

Bug: 1259606
Change-Id: I351a2c8c9054f5b21140a991e5c3760636854488
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3295491
Reviewed-by: Tim van der Lippe <tvanderlippe@chromium.org>
Reviewed-by: Michal Mocny <mmocny@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>