[webauthn] Fix timing issue when WebAuthnPane#model is initialized
Removing the module.json for the legacy DataGrid changed the time
when WebAuthnPane gets instantiated. If the pane is already visible
when DevTools boots up, then the WebAuthnPane constructor runs before
the main target is attached and ready. This causes the model to not
be initialized and a lot of the functionality of the pane will be
missing.

This CL fixes this by properly going through the SDKModelObserver.
The SDKModelObserver is timing independent and guarantees a call
to `modelAdded` for all targets (even if the target already exists).

R=tvanderlippe@chromium.org

Fixed: 1273773
Change-Id: I8b47184eb653941b1d6ddf2c09a611e5f814ea98
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3302734
Auto-Submit: Simon Zünd <szuend@chromium.org>
Commit-Queue: Tim van der Lippe <tvanderlippe@chromium.org>
Reviewed-by: Tim van der Lippe <tvanderlippe@chromium.org>