Fix animation preview popover not showing
Sometimes the animation preview popover does not show up, because when
the mousemove event is fired on the parent element of the preview box,
the check would not pass. This CL fixes this problem by checking both
the event.target element as well as its parent.

After this fix, there is still a very small chance that the popover
won't show up, but this is due to the underministic event order fired
to the popover helper (mousemove to another preview vs. mouseout from
the previous preview). This is WAI for now.

Fixed: 1236943
Change-Id: Ib3c921792cae1956d8a9b2b0b33f59969a01786f
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3302737
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Auto-Submit: Changhao Han <changhaohan@chromium.org>