Add `title` attribute to perf panel time cell
This CL adds the title attribute to the cells showing total time in the
event log breakdown in the performance panel. We show the time taken for
a task in ms but often it's so small that it gets rendered to 1 decimal
place as `0.0ms` which is confusing. Whilst it's not worth the visual
noise of adding more units, we can at least add the title with more
decimal places so users can see just how small a task is.

Bug: 1269109
Change-Id: Ie333afbf1bde306ca96cb2c6b8dd26f995ba4745
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3306502
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Changhao Han <changhaohan@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Changhao Han <changhaohan@chromium.org>