Remove entrypoints/node_app/node_app.json
This CL does a couple of things to preserve the current behavior:
- It folds `node_main` into `node_app`, since we no longer require that
distinction (this was only required prior where we used `autostart` for
`node_main`, which we no longer do)
- It removes as much side-effect instantiation from various modules and
moves them into `node_app.ts`. The order is a bit awkward, but the calls
at the end of `node_app.ts` is the order of side-effect calls as it
exists currently. After we cleaned up all entrypoints, we can remove
some of the calls and clean it up further, but let's keep them here
explicitly to ensure we retain existing behavior.
- Folds `node_app-meta.ts` into `node_app.ts` for simplicity
- We now roll up everything into `node_app.ts` directly, rather than
having a separate meta script. Since this is an entrypoint, we can do
that here. Since the entrypoint will always eagerly load, we don't need
an explicit difference between the meta file and the module
implementation.

R=jacktfranklin@chromium.org

Bug: 1127902
Change-Id: Ida1cee76ae04b3d590dcd16d0225f65852bce5a5
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3306499
Commit-Queue: Tim Van der Lippe <tvanderlippe@chromium.org>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>