Reland "Batch rendering commands and skip rendering if commands are same"
This is a reland of 2288c1657cdb57590a3a6c8c50d7aece4eaaa531 + a fix
for the issue when hiding and showing the same overlay lead to the
overlay not being painted.

Note: on my CPU this optimization reduces cpu usage from ~60% to ~30%
when inspecting a single element that does not change.

Original change's description:
> Batch rendering commands and skip rendering if commands are same
>
> Bug: 1062574
> Change-Id: I2bd2d02055f36a318ecd56194817fb531eb614b9
> Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3301392
> Reviewed-by: Simon Zünd <szuend@chromium.org>
> Commit-Queue: Alex Rudenko <alexrudenko@chromium.org>

Bug: 1062574
Change-Id: I82e6ed03cd7c44f0e8ac7fd78f3623bed8e23187
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3308808
Reviewed-by: Simon Zünd <szuend@chromium.org>
Commit-Queue: Alex Rudenko <alexrudenko@chromium.org>