Fix styling of network waterfall
This particular popover was pulling in content of the other view, but
wasn't properly adding the accompanying styling. Since this is a not a
self-contained component, we need to duplicate the injection of styles
in the other shadowRoot as well.

R=jacktfranklin@chromium.org

Fixed: 1275501
Change-Id: I72a5eeb0902132e8c7402b20beb7e52423ab1985
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3310862
Auto-Submit: Tim Van der Lippe <tvanderlippe@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>