[bfcache] Update the string for OptInUnloadHeaderNotPresent
This CL updates the explanation string for OptInUnloadHeaderNotPresent.
This is an experimental reason and we are not going to launch this. In
order not to confuse the developers, we should just say "The page has
an unload handler in the main frame", same as when unload handler
exists.

Bug: 1275813
Change-Id: Ifde0c287b0b6106d3ecb3db6b67d09dcd2e9c2d2
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3312443
Commit-Queue: Yuzu Saijo <yuzus@chromium.org>
Reviewed-by: Minoru Chikamune <chikamune@chromium.org>
Reviewed-by: Wolfgang Beyer <wolfi@chromium.org>