Add caching to ThemeSupport.getComputedValue
We have support for retrieving computed styles via ThemeSupport. This,
however, does not currently cache values, which in turn means that we
may trigger RecalcStyles on the DevTools UI unnecessarily when querying
these values.

This CL adds a Map to the computed style query. As it stands we never
clear this cache directly as we reload when the theme changes. In
future, if we remove the reload behaviour, we will need to clear this
cache, too.

R=jacktfranklin@chromium.org

Bug: 1278738
Change-Id: I433fae8ef04122b662b9cc767a671f54bd605a50
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3329609
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Paul Lewis <aerotwist@chromium.org>