Add per-target caching to theme support
In the caching behavior I added we assume that we're going to always
query the styles / custom properties from the document. However, there
are times we'll want to query values at the shadow root level. This CL
adds an optional target variable from which we can query the styles, as
well as per-target caching of any values. If the value is empty we don't
cache it, as it likely indicates that the value wasn't ready yet for
some reason, and therefore not worth caching.

R=jacktfranklin@chromium.org

Bug: 1278738
Change-Id: I4c9c8dc54b7ef269e22521b0098982350ea8fb57
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3344908
Auto-Submit: Paul Lewis <aerotwist@chromium.org>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Paul Lewis <aerotwist@chromium.org>