[cleanup] Simplify `NumberUtilities.bytesToString()`.
This removes the unnecessary dependency of `bytesToString()` on
`StringUtilities` and instead implements the logic directly in
terms of vanilla JavaScript. This was discovered as part of the
preparatory work to correctly support the Console Standard in
the DevTools front-end.

Bug: chromium:1282076
Change-Id: Ia1819182abb853eac5cca16e3ffc11194162f511
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3352219
Auto-Submit: Benedikt Meurer <bmeurer@chromium.org>
Reviewed-by: Yang Guo <yangguo@chromium.org>
Commit-Queue: Yang Guo <yangguo@chromium.org>