Make auto dark setting enable `prefers-color-scheme: dark`
The auto dark mode is intended to always be used together with
`prefers-color-scheme: dark`. This change makes the setting in the
Renderer tab both enable auto dark and sets the prefered color scheme to
dark. To make it clear that these two settings are intertwined, we
disable the select box for prefers-color-scheme when auto dark mode is
enabled.
This change also makes the auto dark setting to a checkbox to
simplify the UX.

https: //imgur.com/a/79wNwc0

Change-Id: I3917356865203b89196402a8a7d4dbfa6cbeab18
Bug: none
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3358535
Reviewed-by: Changhao Han <changhaohan@chromium.org>
Commit-Queue: Johan Bay <jobay@chromium.org>