Fix text indent after break for html element in DOM tree
When we compute the indent in the DOM tree, we account for the expand
triangle icon (styled by a ::before rule). However, we consider the
top-level html element to be non-collapsible, and hide the triangle
icon.

That causes the indentation in case of line break on the html tag to be
incorrect and off by the icon's width.

The solution is to not have the expand icon for non-collapsible
(always-parent) elements in the first place.

Fixed:1283369
Change-Id: I1fad80a998b2dfa6f144046040ba04799e01ada1
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3361044
Auto-Submit: Yang Guo <yangguo@chromium.org>
Reviewed-by: Mathias Bynens <mathias@chromium.org>
Commit-Queue: Yang Guo <yangguo@chromium.org>