Restructure scheduleUpdateInDebugger to allow for awaiting it
This CL restructures scheduleUpdateInDebugger to allow callers
to await the update. In follow up changes we will
need to properly await scheduleUpdateInDebugger to make sure
that an update (setting the breakpoint) has happened before
we can continue execution. Specifically, this is relevant
in scenarios in which we stop execution (e.g. through
instrumentation breakpoints) to make sure that all breakpoints
are set, before allowing the execution to continue.

Bug: chromium:1133307, chromium:1229541
Change-Id: I3ef2103ff03aaa02503a680b190c2d46148448de
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3361035
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>
Reviewed-by: Jaroslav Sevcik <jarin@chromium.org>
Commit-Queue: Kim-Anh Tran <kimanh@chromium.org>