Fix GN dependencies
The formatter model creates a worker that references the formatter_worker
entrypoint. However, it wasn't depending on it as such. Therefore,
to ensure deterministic building, we have to add the entrypoint as
dependency to the model. That required another tidy up, as there was
a type defined in the model that was used in the entrypoint. We can
move that type to the entrypoint to circumvent the issue.

Additionally, fix the dependency on Images and themeColors.css to make
them available as expected when running unit tests.

R=jacktfranklin@chromium.org

Fixed: 1284865
Change-Id: I9d1eb6877629ccb5dd5ed105a10acf6291275610
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3369364
Commit-Queue: Tim Van der Lippe <tvanderlippe@chromium.org>
Auto-Submit: Tim Van der Lippe <tvanderlippe@chromium.org>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>