Remove reload when changing themes
This is a fairly widesweeping set of small changes necessary to change
themes dynamically. It mostly involve a change or two to the Theme
Support implementation itself, such that it detects the move from light
to dark theme and, in so doing, clears out its caches and dispatches an
event. This event is used in various places to re-query theme-centric
variables. This applies particularly to the perf UI / flame chart.

Other smaller changes to function signatures are implied by some of the
changes, hence the number of files affected by this CL.

R=jacktfranklin@chromium.org

Bug: 1278738
Change-Id: Iacde026ee69020557491817c8a700b2a7fd2217f
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3369366
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Paul Lewis <aerotwist@chromium.org>