Move normalizePath to ParsedURL
This static function on runtime was only used in ParsedURL, so let's
move it there. Additionally, there is a second definition of this
method (but slightly more simplistic) in the ExtensionServer. Therefore,
let's port that over to the ParsedURL implementation as well.

R=bmeurer@chromium.org

Bug: none
Change-Id: Ie2518e1ff46ba05fc4afd7e4f2ef3f83f2e6a83b
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3369370
Auto-Submit: Tim Van der Lippe <tvanderlippe@chromium.org>
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>
Commit-Queue: Benedikt Meurer <bmeurer@chromium.org>