Fix race conditions in unit tests initialization
There were two problems:
1. We were accidentally using top level await in the EnvironmentHelpers
file. This meant that any import of this file would wait for the promise
of the module to resolve, before it would continue executing the unit
test file. This is problematic, as it means that test definitions aren't
loaded when the file is loaded, but only after the eventual promise
resolves. This problem manifests when you attempt to `it.only` a test,
but Mocha then runs more tests that are discovered after the fact.
2. The initialization bits in `describeWithConnection` weren't properly
awaited. This was caught by the typescript-eslint promise race rule that
we enabled last week, but it was suppressed by using void. By properly
making the functions async and awaiting the results, we resolve these
problems.

Because of these fixes, Jack now owes me a beer.

R=jacktfranklin@chromium.org

Bug: none
Change-Id: I388b8c21e156ddbc85d8d01793a82dede6ea61dc
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3381706
Auto-Submit: Tim Van der Lippe <tvanderlippe@chromium.org>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>