Remove automatic formatting of CSS styles after updating properties
When users interact with CSS properties in the elements panel, they can
enable/disable properties or remove/add properties. In all cases, they
cause the layout engine to apply these changes, while also showing the
differences in the Changes panel.

Historically, we performed the local text replacement and then we
formatted the code using a CodeMirror tokenizer to correct any duplicate
semicolons. The reason that we add several semicolons is in the event
that users declare a property without any semicolons (which is only
legal if it is the last property). You can see this behavior in the
creation of `newStyleText`, which wraps the new `propertyText` in 2
semicolons.

Rather than performing full tokenized formatting using CodeMirror 5, we
can also post-process the text and manually clean up these semicolons.
In this case, we want to replace all duplicate semicolons and remove any
obsolete ones.

As such, we can remove the full CodeMirror TokenizerFactory as defined
in `text_utils`, as this was only used in SDK. This removes the need for
including CodeMirror 5 in our initial load, as we now only use
CodeMirror 5 in our formatter worker. That should be a startup
improvement on its own, while also removing some techdebt.

R=changhaohan@chromium.org

Bug: 1287519
Change-Id: I0d4e92ac9d10de6d15c9ebf05f58ea515773a0f3
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3395675
Reviewed-by: Changhao Han <changhaohan@chromium.org>
Commit-Queue: Tim Van der Lippe <tvanderlippe@chromium.org>