Normalize absolute source URLs in source maps
Update ParsedURL.completeURL to normalize the input URL path even if it
is an absolute URL (i.e. includes a scheme:// component). This will
remove leading ".." segments and resolve any other ".." segments
embedded in the URL. This helps extract neater source URLs from source
maps and as a result, there will be a neater folder tree in the Sources
navigation view that more accurately reflects what the source files
would have looked like on the developers local file system. Mozilla's
source-map npm package already does this when resolving sources.

While testing this solution, I decided to skip appending the
sourceRoot to a source URL if that source URL is absolute. It seems
unlikely that developers would provide both a sourcesRoot and an
absolute source URL, but it's not entirely clear what we should do if
this happens. The source maps v3 proposal doesn't specifically cover
this situation. Naively prepending a non-empty sourceRoot to an absolute
URL probably won't result in a valid or meaningful URL. If the source
URL is already absolute, it seems safe to simply ignore the sourceRoot.

I've added new test cases for ParsedURL and SourceMap to cover the
expected behavior for this change. These all match Mozilla's behavior
except for the cases with a non-empty sourceRoot and an absolute source
URL. Mozilla prepends the sourceRoot to the sourceURL which results in
an invalid final URL and didn't seem worth re-implementing.

Bug: 1284737
Change-Id: If285bd1c16a34bd62e84e547aab75ae48d1063d0
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3362253
Reviewed-by: Tim Van der Lippe <tvanderlippe@chromium.org>
Reviewed-by: Jaroslav Sevcik <jarin@chromium.org>
Commit-Queue: Brandon Walderman <brwalder@microsoft.com>