Show error decorations in editor when snippets throw errors
Snippets have an interesting relationship with UISourceCode. Snippets
are loaded and stored in a "filesystem" project, and have, as such, a
corresponding UISourceCode. When snippets are evaluated however,
DevTools treats them as "normal" page scripts and creates a second
UISourceCode in the "network" project of the main target.

The AutoMapping class is aware of the two snippet UISourceCodes and
connects them together the same way we handle workspaces (where
developers can add a local directory and have it mapped to a pages
resources).

When this mapping happens, the UISourceCodeFrame that displays the
snippet switches modes. UISourceCodeFrame has two modes: It can display
a single UISourceCode, or it can display such a filesystem <-> network
mapped pair of UISourceCodes (aka PersistenceBinding). The bug which
this CL fixes occurs during this switch.

Each UISourceCode has a list of messages attached that serve as the
basis for decorations in the editor. Previously, when the switch from
single UISourceCode to PersistenceBinding happened,
the UISourceCodeFrame would ignore the messages stored on the
newly bound network UISourceCode. It would keep showing the messages
from the filesystem UISourceCode. This CL fixes this by resetting
the rendered decorations to the union of both message sets.

Fixed: 1277403
Change-Id: Ibffb86afcbd1c6a5a5c629a6f9b3b2dff73ab9ed
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3412774
Reviewed-by: Jaroslav Sevcik <jarin@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>