Fix breadcrumbs test
This test relies on the i18n environment to be setup, but it wasn't as
part of the breadcrumbs tests. They worked fine when the entire suite is
run, presumably because another test sets up the environment in time,
but in isolation they fail. This CL fixes that.

Bug: none
Change-Id: I443de4e40b1e238c86d1d55316833e7f83d6319f
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3423896
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Maksim Sadym <sadym@chromium.org>
Commit-Queue: Maksim Sadym <sadym@chromium.org>