[cleanup] Turn 'Project' into a true interface
This CL changes 'Project' from an abstract class into an interface
and makes 'ProjectStore' an abstract base class instead. This removes
the awkward usage of an abstract class in an "implements" clause as
well as the awkward cast of 'this' to 'Project' inside 'ProjectStore'.

R=kimanh@chromium.org

Bug: 1290254
Change-Id: I9a2617557dc02743f4fe0c5c965dc9ea6bf7e675
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3423904
Reviewed-by: Kim-Anh Tran <kimanh@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>