Update scope variable name on source map change
This patch makes the scope chain pane listen to ithe source map attach
event so that it can update itself when a source map is added or
replaced.  This also fixes a race where scope variable names are not
updated if the source map is fully loaded only after the first
breakpoint is hit.

As part of this patch we also started tracking source maps in the scope
resolution cache so that we can ignore the cached scope if it was
computed against an old source map.

The E2E test here adds a source map to a script without a source map and
checks that the scopes view is eventually updated according to the
source map.

Bug: chromium:1294682
Change-Id: Ia2b091ebbc0d50520f0388f35b0d727821bce0d1
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3439918
Auto-Submit: Jaroslav Sevcik <jarin@chromium.org>
Reviewed-by: Kim-Anh Tran <kimanh@chromium.org>
Commit-Queue: Kim-Anh Tran <kimanh@chromium.org>