[sources] Add e2e test for setting breakpoints in live edited scripts
This CL adds a test where we:
  1) Edit a page script
  2) Set a breakpoint in the edited script
  3) Reload the page (this reverts the script)

Ideally the breakpoint moves in the reverted script in a smart way.
We have some infrastructure for that in place on the V8 side, but the
DevTools frontend does not account for it correctly.

Once the bug is fixed we can update the test expectations accordingly.
Rather than landing a skipped test, we expect the bugged state so we
get notified if the test starts breaking for unrelated reaons.

R=kimanh@chromium.org

Bug: 1216904
Change-Id: Ife4f4538b216c42888013363ef5e6cefb9b1392a
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3447410
Reviewed-by: Kim-Anh Tran <kimanh@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>