Fix incorrect repositioning of CodeMirror editors after reloading
When reloading an editor, we restore the original position that the
editor was in. However, with CodeMirror 6, we were using an effect
that implicitly was centering on the line, whereas the internal
DevTools history would store the first line of the editor. As such,
when you would reload the editor multiple times, it would jump
up a couple of times.

Since this effect is also deprecated and replaced with `scrollIntoView`,
we should use that instead. With that new method, we can also specify
the "y"-behavior to "start", rather than the default "center", to get
the correct repositioning behavior.

R=yangguo@chromium.org

Fixed: 1294422
Change-Id: I11ba56a011d1c293c35d0d0a552231b107ac372d
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3447413
Auto-Submit: Tim Van der Lippe <tvanderlippe@chromium.org>
Reviewed-by: Yang Guo <yangguo@chromium.org>
Commit-Queue: Yang Guo <yangguo@chromium.org>