[console] Always append the source location for SyntaxErrors
This CL changes how we render SyntaxErrors in the DevTools console.
SyntaxErrors don't contain a stack frame with the source position
where exactly the SyntaxError is. This is because V8 throws
SyntaxErrors while parsing, not at runtime. At most, the top-most
stack frame is a call to 'eval'.

This CL uses the info from `Runtime.getExceptionDetails` to append
the source location of the error to the console message.

Screenshot: https://imgur.com/a/x5HP11q

Please note that the Runtime.getExceptionDetails info is only
available when DevTools was already open when the script fails
to parse. This is because the inspector does not cache failed
parse events (opposed to successful script parses).

R=bmeurer@chromium.org

Bug: 1066909
Change-Id: Ibf013d978c13970a688326d9afb9bf0bb14eff99
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3461550
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>