Stack Location: Backwards focus errors
In Stack Location views, the title was marked as a `button` for ARIA
purposes. When the title contained a toolbar which itself contained a
button, the screen readers would fail to announce the title of the view
when the user used keyboard navigation to go backwards (e.g., Shift+Tab)
ostensibly because it was the transition of button to button.

Per MDN:

https://developer.mozilla.org/en-US/docs/Web/Accessibility/ARIA/Roles/tab_role

The 'tab' role is for an element which controls the visibility of
another element. The title bar of stack locations sufficiently applies
here, and in fact, the title bar already receives the `aria-controls`
attribute.

Bug: 1166709
Change-Id: Ie71b6814e826d6f968afb874e2a9f751723ff283
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3461315
Reviewed-by: Michael Liao <michael.liao@microsoft.com>
Reviewed-by: Mathias Bynens <mathias@chromium.org>
Commit-Queue: Robert Paveza <Rob.Paveza@microsoft.com>