Hide color picker input from tab navigation
The implementation of the color picker for Element panel's Layout pane
is a bit unusual. The label acts as input element, receiving the event
to open the color picker. The actual color input is hidden, but still
can be focused through keyboard navigation (tab).

That causes an exception when focusing the input element and pressing
space, because the onColorLabelKeyUp event handler does not expect the
input element as event target. It also causes the user having to tab
twice to arrive at the next visible UI element.

R=changhaohan@chromium.org

Fixed: 1292969
Change-Id: Id6f7d0e3d8b330088f98cadf0379496a0d636d6c
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3473367
Reviewed-by: Changhao Han <changhaohan@chromium.org>
Commit-Queue: Yang Guo <yangguo@chromium.org>