[objects] Fix revealing private accessor values.
Previously, private accessors couldn't be revealed by clicking on the
(...) because we'd try to literally evaluate `obj["#field"]`, which
looks for the named property "#field" rather than a private field. This
fixes the objects display (and in particular the Scope view) to instead
invoke the getter function directly.

Fixed: chromium:1296855
Change-Id: I567d498229d616f905c285a8a541e50b490ca567
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3479594
Reviewed-by: Kim-Anh Tran <kimanh@chromium.org>
Commit-Queue: Benedikt Meurer <bmeurer@chromium.org>