Improve handling of file name changes for local overrides
One can change the file type of a file in the local overrides folder by
renaming the file. Such a rename can cause a different viewer to be
more appropriate for handling this file (e.g. when renaming a html-file
into a jpg, the file should now be handled by an image viewer). This CL
makes sure that in this scenario, the existing editor tab is closed and
a new one with an editor of the correct type is opened.

Unsaved changes in the file are not lost and it is ensured that there
is no unnecessary dialog popping up with a warning about external
changes to the file.

Bug: 1297532
Change-Id: I95098e15455c2ca328be00515941fe0950e5e3e0
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3468521
Reviewed-by: Danil Somsikov <dsv@chromium.org>
Commit-Queue: Wolfgang Beyer <wolfi@chromium.org>