[DOMModel] Request document if it hasn't happened yet
The documentUpdated() function is called upon a DOM.documentUpdated
CDP event. Before, we would *always* set the document to null, but
only request the document if *we previously had a document* or
we *already requested* a document. This changes the logic to always
set it to null and make sure to request an update if there is not
already an ongoing update.

Bug: none
Change-Id: I97604afbe7904970a08a54c5b6ef969e12c64963
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3477101
Reviewed-by: Simon Zünd <szuend@chromium.org>
Commit-Queue: Kim-Anh Tran <kimanh@chromium.org>