Add back console autocompletion for map.get
With the upgrade to CodeMirror 6, we didn't include the autocompletion
for maps yet. This CL adds back for detection of `map.get(` which will
list all available keys on the map.

If the object is not a map, then the `evaluateExpression` will contain a
`result.exceptionDetails`, which means we gracefully ignore the result.
However, if it is map, we traverse the array and obtain all keys (which
are strings). The autocompletion will then also include the closing
paraenthesis, so that the result would become:

```js
const map = new Map([["somekey", 5], ["some other key"]]);
map.get(<auto-complete><press tab>
map.get("somekey")
```

The keys of the map are ordered based on the original appearance in the
keys array to ensure a deterministic autocompletion value.

Corresponding unit and e2e tests are written that verify the user story.

R=bmeurer@chromium.org

Bug: 1293807
Change-Id: I53dd9559e4a7e13fe4e36e934b66a693f606ec86
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3497704
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>
Commit-Queue: Tim Van der Lippe <tvanderlippe@chromium.org>