Fix emulation toolbar not respecting theme
When we did the final migration to remove the requirement to reload
DevTools on a theme change, this was missed and as a result the
emulation bar was always rendering in the light theme.

With this change it will now respect the current theme and update live
if the user changes theme.

Fixed: 1304243
Change-Id: I37fc7b255599f1eed29219729f07d6d024dbbf7b
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3525355
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Mathias Bynens <mathias@chromium.org>
Commit-Queue: Mathias Bynens <mathias@chromium.org>