Fixes mixed-sortable data grid click-to-sort.
The underlying cause is that, when there are no sortable columns, the
header row isn't selectable at all. However, when some columns are
sortable, the handling of clicking is passed through indiscriminately.

This change checks the sortability of the column before passing the
event through to the sort handler.

Bug: 1310818
Change-Id: I4326df5a0c55ecfba88be0f5586ee8e5c2c0f0b6
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3556086
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Robert Paveza <Rob.Paveza@microsoft.com>