Correctly support multiple slashes in URLs.
According to RFC 2396 as well as RFC 3986 every slash in a URL is
significant, and especially in case of HTTP, it's up to the server
to interpret these. So we must not strip redundant slashes in the
client, in particular not in the case of URLs coming from source
maps.

Fixed: chromium:1310746
Bug: chromium:1284737
Change-Id: Iec13e6647cacc812af86a4b5904b207ee3c94f18
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3555677
Reviewed-by: Jaroslav Sevcik <jarin@chromium.org>
Commit-Queue: Benedikt Meurer <bmeurer@chromium.org>
Auto-Submit: Benedikt Meurer <bmeurer@chromium.org>