Fix normalization for URLs that contain optional user part.
When passing a URL as href to `ParsedURL.completeURL()`, we'd take it
apart to normalize the path and strip off user and password parts. But
in those cases where the URL contained the optional user part, we didn't
do the math correctly for reconstructing the query and fragment part,
and instead ended up duplicating the tail of the original href. This CL
addresses the issues and guarantees that we always reconstruct a valid
URL.

Note: The stripping of the optional user part is consistent with Firefox
DevTools. So 'webpack://@scope/...' turns into 'webpack://scope/...' in
both DevTools consistently.

Fixed: chromium:1309854
Bug: chromium:1284737
Change-Id: I45224265e7d6cc83f836fafc516e71434e9ba402
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3560422
Auto-Submit: Benedikt Meurer <bmeurer@chromium.org>
Reviewed-by: Yang Guo <yangguo@chromium.org>
Reviewed-by: Brandon Walderman <brwalder@microsoft.com>
Commit-Queue: Benedikt Meurer <bmeurer@chromium.org>