[A11y] Screenreader announcing incorrect row count for table
ScreenReader is announcing incorrect row count, which is caused by the filler rows in the DataGrid.
This CL remove unnecessary tabIndex="-1" and set the filler rows to aria-hidden="true" to completely hide them from ARIA APi. (tabIndex="-1" is needed when we want to focus on the element programmatically, which is not the case here)


Bug: 1313315
Change-Id: I74d14b3343002d2b18e0c504f6b0a462c91dabcf
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3570810
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Christy Chen <chrche@microsoft.com>