Reduce top margin on PanelIntroductionSteps
It was pushing all the introduction views down slightly. As the <h1>
is at the very top of the component we set the margin to 0 so whoever is
rendering this component can apply their own margins if they require.

Bug: 1311259
Change-Id: I3b0c4fcd988ebb503f3a29ff67e0b96dd8ed21dc
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3574566
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Changhao Han <changhaohan@chromium.org>
Commit-Queue: Changhao Han <changhaohan@chromium.org>