Always force white icons for selected items in a soft menu.
The if statement seem to originate in the crrev.com/c/1161282 which
introduced light gray highlight for the selected menu item on win and
linux. As we are back to blue hightlight everywhere we also need to make
icons white everywhere.

Bug: 1309777
Change-Id: I1d2b053aea2b07103ea6e42c6bb701b4c3be88a3
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3586911
Auto-Submit: Danil Somsikov <dsv@chromium.org>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>