Mark inherited and overridden rules for highlights
As a follow-up to [1], this change makes two updates for inherited
highlight pseudos:

- Draw a strikethrough through overridden styles.
- Don't grey-out non-inherited properties in the context of highlight
pseudos, since the highlight cascade treats all valid properties as if
they were inherited [2].

See [3] for screenshots of the change.

For now, custom highlights are not included in this change. Custom
highlights are currently grouped together into the same
`DOMInheritanceCascade` regardless of highlight name, so custom
highlight styles could be shown as overridden by styles from a
highlight with a different name. To fix this, custom highlights
would need to be divided into separate inheritance cascades.
This is nontrivial, so that work will be split up into a separate
change.

[1] https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3502175
[2] https://www.w3.org/TR/css-pseudo-4/#highlight-cascade
[3] https://docs.google.com/document/d/1hU-KL5-WIbTWSH46N7xaP8SBm69QAMiyg5jOY7ZHVTU/edit

Bug: 1024156
Change-Id: Ib32d862bba1ce1843b55d3e85f0e6e9f3a2447e3
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3563627
Reviewed-by: Delan Azabani <dazabani@igalia.com>
Reviewed-by: Johan Bay <jobay@chromium.org>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Commit-Queue: Dan Clark <daniec@microsoft.com>