Fix local overrides for redirects
When DevTools is intercepting and overriding the response to a request,
this response should have a response status of 200 (and not the response
status of the original response).

If body, headers and a redirect response code are sent to the backend
via `Fetch.fulfillRequest`, the response code will take precedence and
the response will be redirected instead of overridden.

Fixed: 1320400
Change-Id: Ib306369c67471f871425d2a1bbafe80cbb881d1f
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3610172
Commit-Queue: Wolfgang Beyer <wolfi@chromium.org>
Reviewed-by: Danil Somsikov <dsv@chromium.org>
(cherry picked from commit d939ee8ddb8e751116fa9fdb702adb8c0710113b)
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3627334
Bot-Commit: Rubber Stamper <rubber-stamper@appspot.gserviceaccount.com>