[e2e] For debugging, pause when tabs are ready
Enabling the pool delayed tab creation into the `beforeEach` block.
So when we pause in the `beforeAll` block in debug mode, the DevTools
and target tab are not yet booted up.

This CL moves the pause into the `beforeEach` block but still only
pause once.

R=pfaffe@chromium.org

Bug: 1297458
Change-Id: Ic3f1ddc41696d4e11c6797effc20d1101399d163
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3642145
Reviewed-by: Philip Pfaffe <pfaffe@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>