[l10n] De-duplicate deprecation message UI strings
Locked terms are implemented as placeholders that are always replaced
with the same value. That means that many of the UI strings in
DeprecationIssue.ts are currently duplicates.

The TC does not like importing duplicated strings, so we de-duplicate
them in this CL.

R=kimanh@chromium.org

Bug: None
Change-Id: I79997be64c57107f34622b808b9a9adff34123ec
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3642144
Commit-Queue: Simon Zünd <szuend@chromium.org>
Auto-Submit: Simon Zünd <szuend@chromium.org>
Reviewed-by: Kim-Anh Tran <kimanh@chromium.org>