[debugger] Implement 'Restart frame' context menu item
This CL adds back the "Restart frame" context menu item that was
previously removed. The difference is that now we enable/disable
the menu item based on the `canBeRestarted` flag we get for each
call frame in the Debugger.paused event.

Note that the feature itself is mainly tested in inspector tests. We
only add a single e2e test to check that the core user story works.

R=kimanh@chromium.org

Bug: 1303521
Change-Id: I47890af743482fcf53de4a1a6a6474efb9a1da56
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3652197
Commit-Queue: Simon Zünd <szuend@chromium.org>
Reviewed-by: Kim-Anh Tran <kimanh@chromium.org>