[A11y] Fix DOMBreakpoint sidebar screen reader issue
Screen Reader is announcing the DomBP content twice and not announcing the checkbox state in DOMbp sidebar. This is because the focusable element is the list element, but the element itself doesnt have an aria-label, so the screen reader will read through the label of the checkbox and the BP description, causing it to announce it twice.
Now add the aria-label with complete info to the list element so when the bp item get focused, the screen reader would read out the complete info once.

Bug:1327440
Change-Id: I94fd9d3e0b93377f684fea6b5d384c944e29dc3d
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3654682
Commit-Queue: Christy Chen <chrche@microsoft.com>
Reviewed-by: Mathias Bynens <mathias@chromium.org>
Reviewed-by: Johan Bay <jobay@chromium.org>