[e2e] Fix instrumentation test for inline scripts with sourceURL comment
This test fixes the test for instrumentation breakpoints on
inline scripts with sourceURL comment. A missing space in the
source url comment of the input html made that it wasn't recognized
as such. This CL now fixes the typo and changes the test to reflect
restoration of breakpoints on scripts with sourceURL comment.

Bug: 1229541
Change-Id: I3299983f2607112cef08defd810c92009994e1c2
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3675716
Commit-Queue: Kim-Anh Tran <kimanh@chromium.org>
Reviewed-by: Jaroslav Sevcik <jarin@chromium.org>