Introduce explicit opening/closing tag types in DOM tree outline
We currently have a few runtime type errors from trying to add an
adorner to a `close` html tag element.
This change introduces a more strict TS typing for DOM tree elements
fields that are only applicable for opening tags. This more explicit
typing ensures that we can't accidentally get into an inconsistent state
where only some of the necessary fields are populated.
As part of this work we also remove a few `@ts-expect-error`.

Bug: 1172300
Change-Id: Iaf80517a76b6bd7e8edee115aad441e5101fd19b
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3675715
Commit-Queue: Johan Bay <jobay@chromium.org>
Reviewed-by: Changhao Han <changhaohan@chromium.org>