Changed aria control type for ViewManager elements
Modify ViewManager so that StackLocation/ExpandableContainer widgets are
correctly read by screen readers.

Previously they were declared as 'role: tab' which meant that
tools would not read the 'aria-expanded' label since tabs do not usually
have expanding behavior. This change marks the containing StackLocation
as a tree, which allows for treeitem's underneath it to have expand/collapse labels that are correctly read.

This StackLocation element is used in 2 places, the right panel in sources
view and the accessibility sidebar view. Since both of these locations had the
incorrect behavior before, this seemed like the correct scope for this change.

Change-Id: If90253640fb8b00c3c2f7932e19a198d216c0b2c
Bug: none
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3673131
Reviewed-by: Simon Zünd <szuend@chromium.org>
Commit-Queue: Jake Mcleman <jamclema@microsoft.com>