Added aria-labels to DataGrid container
DataGrids already have aria-labels to convey their width and height but
not a label to let screen reader dependent users know which table they
have selected. This is cumbersome on pages where there is only one
DataGrid as the user has to move between the title and the table to
find all of the information about it, but makes panels with more than
one DataGrid hard for them to understand.

This change adds a field on DataGrid and DataGridController to allow
for a label to optionally provided, and if it is provided an aria-label
tag will be applied to the table.

Change-Id: Ie6c9b0de64e120ef9685cfa0141a7081a5c5abb3
Bug: none
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3673135
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Jake Mcleman <jamclema@microsoft.com>