typescript: Rename (devtools_)use_rbe to (devtools_)use_remoteexec
It is safe to assume that use_remoteexec is set if use_rbe is set: https://source.chromium.org/chromium/chromium/src/+/main:build/toolchain/rbe.gni;l=29-33;drc=5387d0ef00a01a27a1cd5b68a5612485f88830f7

The use_rbe argument is used to enable the reclient remote execution
client, which is not the same as the remote build execution service.
Re-client is used to communicate with RBE, but it is not the same thing.

Bug: 1247781
Change-Id: I4bc64201bab998146727296d37666c0926ce161f
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3692626
Reviewed-by: Takuto Ikuta <tikuta@chromium.org>
Reviewed-by: Simon Zünd <szuend@chromium.org>
Commit-Queue: Richard Wang <richardwa@google.com>