Await network persistence updates on instrumentation breaks
The NetworkPersistenceManager maps UISourceCodes from overrides to their
corresponding network UISourceCodes. In order to make sure that
breakpoints set on the file system are moved to the network one,
we need to wait for the NetworkPersistenceManager to create the binding,
and the PersistenceImpl to move the breakpoint.

Bug: 1280621
Change-Id: I0fee0eb7f115cb2e85aff8f3b4c55aca7a9ca683
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3675720
Reviewed-by: Jaroslav Sevcik <jarin@chromium.org>
Commit-Queue: Kim-Anh Tran <kimanh@chromium.org>
Reviewed-by: Simon Zünd <szuend@chromium.org>