Expose fontDisplay from CSSFontFace
This exposes the fontDisplay property which was recently added to the
Protocol.CSS.FontFace type in the frontend representation of a FontFace.

This doesn't have any functional impact yet, but it will useful when
in need to get the 'font-display' property of a loaded font face.

Bug: none
Change-Id: Ifd288e0f38caa110c7bef40163304586ef8e9108
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3696157
Auto-Submit: Andres Olivares <andoli@chromium.org>
Reviewed-by: Danil Somsikov <dsv@chromium.org>
Commit-Queue: Danil Somsikov <dsv@chromium.org>