[debugger] Add test for scope view resolution with source map
This adds a test that deobfuscates scope view variables based on a
real source map and a scope that is provided via a protocol message.

As a drive-by, the CL also changes how the identifier source
positions are grouped and handled. We also add a (currently failing)
test for source map resolution that motivates this refactoring.
The actual fix is coming soon.

Bug: chromium:1335338
Change-Id: Ib364a0d1fbcce6c1c9c433aa4c4945895981d501
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3695366
Reviewed-by: Simon Zünd <szuend@chromium.org>
Commit-Queue: Jaroslav Sevcik <jarin@chromium.org>
Reviewed-by: Kim-Anh Tran <kimanh@chromium.org>