[liveedit] Promisify Script#editSource
This CL changes Script#editSource to return the result via its Promise
instead of a dedicated callback. This CL also removes the indirection
via the DebuggerModel, which was a bit weird: We had the Script object
already in ResourceMapping, but would only use the scriptId and then
the DebuggerModel would resolve the script object again.

A follow-up CL will change the `error` string to use an enum that
incorporates the new LiveEdit response status as well to improve
error reporting.

R=kimanh@chromium.org

Bug: 1334484
Change-Id: I654fb94a1969e2b56b2e222d6e1c8f495d53b9bb
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3711560
Reviewed-by: Kim-Anh Tran <kimanh@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>