[liveedit] Allow editing of the top-most function on the stack
This CL enables the new flag on the Debugger#setScriptSource
CDP method. This allows editing of the top-most function on the stack
while paused, if that activation is the only activation record of
the edited function on the stack.

In contrast to the old implementation we do not have to explicitly
trigger a "Step In" as the backend automatically restarts the
edited function and the frontend handles the "resumed" + "paused"
events normally.

R=jarin@chromium.org, kimanh@chromium.org

Fixed: 1334484
Change-Id: Ia60ab60646219028123754655b36f27cdff283f3
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3706925
Commit-Queue: Simon Zünd <szuend@chromium.org>
Reviewed-by: Kim-Anh Tran <kimanh@chromium.org>
Reviewed-by: Jaroslav Sevcik <jarin@chromium.org>