[liveedit] Use the new status enum for Debugger#setScriptSource
This CL utilizes the new status enum to decide whether to display an
error or update the script's content.

Keep in mind that until now we didn't actually show the console
message for a failed live edit. That is because all errors were
reported with an "exceptionDetails" object even for non compile
errors.

So already this CL is a great improvement for error reporting
as non-compile live edit failures get finally printed to the
console.

R=jarin@chromium.org, kimanh@chromium.org

Bug: 1334484
Change-Id: I27f484fa510bd2a2f9c43d250f7e9d1fdb581a3e
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3706923
Reviewed-by: Kim-Anh Tran <kimanh@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>
Reviewed-by: Jaroslav Sevcik <jarin@chromium.org>