Link to script which causes a frame to be marked as an ad frame
For frames which are labelled as being an ad, if the backend sends
information about `adScriptId` and `debuggerId`, use this info to create
a link to the script from the frame details view. This link points from
the ad frame to the (bottom-most) script which is responsible for the
frame to be labelled as an ad.

In order to achieve this, we need to
1) keep track of which `debuggerId` corresponds to which `target`. We
do this in `TargetManager`.
2) not lose `adScriptId` and `debuggerId` during a frame's OOPIF
transfer. We take care of this in the `FrameManager`, which already has
been doing analogous work for stack traces.

Screencast: https://i.imgur.com/gF56hEc.mp4

This is a follow-up to https://crrev.com/c/3168916

Bug: 1217041
Change-Id: I99ce83567389f076130d8b48af0f2bb9bf6a3111
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3653565
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>
Commit-Queue: Danil Somsikov <dsv@chromium.org>