Differentiate errors when updating the debugger
Previously `updateInDebugger` would return an error if one occurred.
There may be two sources of errors: 1. a breakpoint clash, 2. a
back-end error. In both cases, we want to remove the breakpoint.

However, in the first case, we also want to remove it from the
storage. Following scenario:

1. User sets breakpoint on line 3
2. User sets breakpoint on line 4
3. Breakpoint clash: breakpoint 4 is removed as it also resolves to
line 3, but remembered in the settings
4. Remove breakpoint on line 3
5. Reload
6. We would hit on line 3, because of the second breakpoint that
was remembered and restored, even though the user only saw one
breakpoint which was removed in step 4.

In order to remove this behavior, we differentiate between the
erros that we can get when updating the breakpoint and act
accordingly (keep in storage if server error, remove from storage
if breakpoint clash).

Bug: 1334930
Change-Id: Id72e137e56dfe78e029cdcd7988e5babfd26f096
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3733114
Reviewed-by: Jaroslav Sevcik <jarin@chromium.org>
Commit-Queue: Kim-Anh Tran <kimanh@chromium.org>