Prepare components for revealing object size
This CL lays the ground work for revealing object size inside of
the Linear Memory Viewer. The changes are limited to associated
components, which leaves the upstream integration flexible.

The change introduces an interface (memoryObjectInfo) for
packaging information related to a memory object.
Currently, that is its size and starting address, and will
likely be extended as necessary.

The transport of memoryObjectInfo is handled between the
LinearMemoryInspector and LinearMemoryViewer. The logic for
highlighting byte-cells and text-cells of the given memory object
is also included.

Bug: 1336568
Change-Id: I750611a297e643fd492e342c89bfc7c81f816bd6
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3733111
Auto-Submit: Michal Pitr <michalpitr@google.com>
Reviewed-by: Kim-Anh Tran <kimanh@chromium.org>
Commit-Queue: Michal Pitr <michalpitr@google.com>