[e2e] Fix performance_profiling_test.ts
This fixes the tests to explicitly await that we have the
correct component before continuing with the tests.

This CL also removes one of the tests that checks if we
can correctly resolve links to the sources tab. It relied on
an uploaded performance trace for that. This is problematic,
as a new run of that page may report different scriptIds than
what was recorded before. Therefore it would in some cases
resolve, and others not.

Bug: none
Change-Id: I14d07ab59f5dd0c0aa07637afeabb1cb2a4e0c2f
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3650752
Commit-Queue: Kim-Anh Tran <kimanh@chromium.org>
Auto-Submit: Kim-Anh Tran <kimanh@chromium.org>
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>