Fix adorner and layout pane update issues
This CL fixes two issues regarding delayed updates of style adorners
and the Layout pane.
For style adorners, the issue is that sometimes
after navigation, the css agent has not been enabled when the frontend
starts to request new styles for nodes. This would result in the backend
returning errors and the frontend not getting correct style.
For the Layout pane, the issue is that the layout pane's Grid and Flex
sections do not update when there are nodes that become or stop being
Grid/Flex containers.
This CL fixes both of them.

Fixed: 1340441, 1273992
Change-Id: I0b86f0bbfa15179e90c345d21abdaa89c0dbd0a1
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3740726
Commit-Queue: Ergün Erdoğmuş <ergunsh@chromium.org>
Auto-Submit: Changhao Han <changhaohan@chromium.org>
Reviewed-by: Ergün Erdoğmuş <ergunsh@chromium.org>