Re-enable data_grid interaction test on Mac.
I have tried without success to recreate this issue on my Mac; I've run
it with drastic slow-down and throttling, and did `it.repeat(100)` more
than once without a single failure.

Given that, and that this test has been in our codebase for a long time
and this is the first time it flaked, I'm inclined to re-enable it for
now without spending more time on it. Should it flake again, we will
disable it and investigate more (hence me using `Bug:`, not `Fixed:`)
here.

Bug: 1340177
Change-Id: I7f47891dd63ec232af160b6cd1fd93b5eb9b1117
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3752796
Commit-Queue: Andres Olivares <andoli@chromium.org>
Reviewed-by: Andres Olivares <andoli@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>